refactor: Core 모듈의 spring boot web관련 기능을 spring-boot-web 모듈로 분리
