+++ core/spring-boot-web/src/main/java/com/github/can019/base/core/spring/boot/web/Main.java
@@ -1,0 +1,7 @@
+package com.github.can019.base.core.spring.boot.web;
+
+public class Main {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
