--- core/src/test/java/com/github/can019/base/core/interceptor/resource/LoggingInterceptorTestController.java
@@ -1,15 +1,0 @@
-package com.github.can019.base.core.interceptor.resource;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-public class LoggingInterceptorTestController {
-
-    @GetMapping("test/logging-interceptor/get")
-    public ResponseEntity getMappingMethod() {
-        return new ResponseEntity(null, HttpStatus.OK);
-    }
-}
